Feat: StateLazyGrid 분리 & 상태 호이스팅
