+++ core/ui/src/main/java/com/kolown/porring/core/ui/component/PageItemFooter.kt
@@ -1,4 +1,4 @@
-package com.kolown.porring.feature.my.component
+package com.kolown.porring.core.ui.component
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
@@ -17,7 +17,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.paging.LoadState
-import com.kolown.porring.feature.my.R
+import com.kolown.porring.core.ui.R
 
 @Composable
 internal fun PageItemFooter(
