+++ feature/main/src/main/java/com/kolown/porring/feature/main/component/MainNavHost.kt
@@ -98,7 +98,6 @@ internal fun MainNavHost(
             theirNavGraph(
                 popBackStack = navigator::popBackStack,
                 navigateToDetailTheir = navigator::navigateToDetailTheir,
-                getBackStackEntry = { navigator.navController.getBackStackEntry(MainMenu.HOME.route) },
                 padding = padding
             )
